Implement exportSecrets
Allows us to set environment variables that could be the named the same between multiple steps, but have different values due to enivronmental differences.

This is a redo of #28 because I'm bad at git. The secrets are exported so that they can be used in any other flag in this plugin.